pyproject.toml
[project]
name = "rssmcp"
version = "0.1.0"
description = "Simple RSS MCP Server"
readme = "README.md"
requires-python = ">=3.10.16"
dependencies = [
 "feedparser>=6.0.11",
 "mcp>=1.6.0",
 "pyyaml>=6.0.2",
 "types-pyyaml>=6.0.12.20250402",
]
[[project.authors]]
name = "Lunran"
email = "feitcounter@gmail.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "build>=1.2.2.post1",
    "hatch>=1.14.1",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.5",
    "twine>=6.1.0",
]

[project.scripts]
rssmcp = "rssmcp:main"
